requirements.txt
absl-py==2.1.0
antlr4-python3-runtime==4.9.3
backcall==0.2.0
beautifulsoup4==4.12.3
cachetools==4.2.4
certifi==2022.12.7
charset-normalizer==3.3.2
datadings==3.4.6
decorator==5.1.1
easydict==1.9
einops==0.6.1
fastrlock==0.8.2
filelock==3.12.2
fsspec==2023.1.0
ftfy==6.1.1
gdown==4.7.3
google-auth==1.35.0
google-auth-oauthlib==0.4.6
grpcio==1.62.0
h5py==3.7.0
huggingface-hub==0.16.4
idna==3.6
imageio==2.31.2
importlib-metadata==6.7.0
ipdb==0.13.3
ipython==7.34.0
jedi==0.19.1
joblib==1.3.2
lmdb==1.0.0
loguru==0.7.2
Markdown==3.4.4
MarkupSafe==2.1.5
matplotlib-inline==0.1.6
msgpack==1.0.5
msgpack-numpy==0.4.8
natsort==8.4.0
networkx==2.6.3
numpy==1.21.6
oauthlib==3.2.2
omegaconf==2.3.0
opencv-python==4.9.0.80
packaging==23.2
parso==0.8.3
pexpect==4.9.0
pickle5==0.0.12
pickleshare==0.7.5
Pillow==9.5.0
prompt-toolkit==3.0.43
protobuf==3.20.0
ptyprocess==0.7.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pyflakes==2.2.0
Pygments==2.17.2
PySocks==1.7.1
PyWavelets==1.3.0
pyzmq==25.1.2
regex==2023.12.25
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
safetensors==0.4.2
scikit-image==0.19.3
scikit-learn==0.24.0
scipy==1.7.3
simplebloom==1.0.5
simplejpeg==1.7.2
six==1.16.0
soupsieve==2.4.1
tensorboard==2.3.0
tensorboard-plugin-wit==1.8.1
threadpoolctl==3.1.0
tifffile==2021.11.2
timm==0.4.12
tqdm==4.46.0
traitlets==5.9.0
typing_extensions==4.7.1
urllib3==2.0.7
wcwidth==0.2.13
Werkzeug==2.2.3
yacs==0.1.6
zipp==3.15.0
